feat: 实现项目生成

+++ server/controller/project.js
@@ -1,4 +1,6 @@
+const {genPageSchemaJson} = require('../utils/download');
 const model = require('../models/project');
+const authModel = require('../models/auth');
 module.exports = {
     getList: async ctx => {
         try {
@@ -9,6 +11,9 @@ module.exports = {
         }
     },
     getProject: async ctx => {
+        if (!ctx.params.id) {
+            throw new Error('项目id为空');
+        }
         try {
             const result = await model.Dao.getProject({userId: ctx.query.user, id: ctx.params.id});
             ctx.body = result;
@@ -25,6 +30,10 @@ module.exports = {
         }
     },
     update: async ctx => {
+        const {id} = ctx.request.body;
+        if (!id) {
+            throw new Error('项目id为空');
+        }
         try {
             const result = await model.Dao.update(ctx.request.body);
             ctx.body = result;
@@ -33,12 +42,36 @@ module.exports = {
         }
     },
     delete: async ctx => {
-        const ids = Array.isArray(ctx.request.body.id) ? ctx.request.body.id : [ctx.request.body.id];
+        const {id} = ctx.request.body;
+        if (!id) {
+            throw new Error('项目id为空');
+        }
+        const ids = Array.isArray(id) ? id : [id];
         try {
             const result = await model.Dao.delete(ids);
             ctx.body = result;
         } catch (err) {
             throw err;
         }
+    },
+    download: async ctx => {
+        const {id} = ctx.request.body;
+        if (!id) {
+            throw new Error('项目id为空');
+        }
+        let project = {pages: '[]'};
+        let user = {};
+        try {
+            // 这里不关联用户，后续可能可以下载别人分享的项目
+            project = await model.Project.findOne({_id: id});
+            user = await authModel.Dao.getUser(ctx.request.body.user);
+        } catch (err) {
+            throw err;
+        }
+        const pages = JSON.parse(project.pages);
+        // console.log(pages);
+        ctx.body = pages;
+        // 生成page json 文件
+        genPageSchemaJson(user?.username, project.path, pages);
     }
 };
